add compile and reorg files
